proyecto_tomate/backend/requirements.txt
flask>=3.0.0
flask-cors>=4.0.0
ultralytics>=8.3.0
tensorflow>=2.20.0
pillow>=10.0.0
werkzeug>=3.0.0
gunicorn>=21.2.0


